aw-transform: Split out transform to a seperate lib
- Split transform to aw-transform lib
- Moved transforms to seperate files
- Moved tests from tests/ to each seperate transform file
