+++ Cargo.toml
@@ -13,6 +13,7 @@ legacy_datastore_tests = []
 members = [
     "aw-client-rust",
     "aw-models",
+    "aw-transform",
 ]
 
 [lib]
@@ -49,6 +50,7 @@ gethostname = "0.2"
 regex = "1.0"
 uuid = { version = "0.8", features = ["serde", "v4"] }
 aw-models = { path = "aw-models" }
+aw-transform = { path = "aw-transform" }
 
 [target.'cfg(target_os="android")'.dependencies]
 jni = { version = "0.5", default-features = false }
