+++ aw-transform/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "aw-transform"
+version = "0.1.0"
+authors = ["Johan Bjäreholt <johan@bjareho.lt>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+serde = "1.0"
+rocket = "0.4" # TODO: Remove this dependency
+serde_json = "1.0"
+regex = "1.0"
+log = "0.4"
+chrono = { version = "0.4", features = ["serde"] }
+aw-models = { path = "../aw-models" }
