+++ src/lib.rs
@@ -29,7 +29,6 @@ extern crate fern;
 extern crate toml;
 
 #[macro_use] pub mod macros;
-pub mod transform;
 pub mod datastore;
 pub mod query;
 pub mod endpoints;
@@ -43,3 +42,4 @@ pub mod android;
 pub mod sync;
 
 extern crate aw_models;
+extern crate aw_transform;
